made task

+++ BottleOfWater.java
@@ -1,21 +1,21 @@
 public class BottleOfWater extends Product{
-    private Double volume; 
+    private int volume; 
     
-    public BottleOfWater( String name, String brand, Double volume, Double price){
+    public BottleOfWater( String name, String brand, int volume, Double price){
         super(name, brand, price);
         this.volume = volume;
     }
    
     @Override
     String displayInfo() {
-        return String.format("%s от %s - Объём %.1f л. - Цена %.2f руб.", name, brand, volume, price);
+        return String.format("%s от %s - Объём %d л. - Цена %.2f руб.", name, brand, volume, price);
     }
 
-    public Double getVolume() {
+    public int getVolume() {
         return volume;
     }
 
-    public void setVolume(Double volume) {
+    public void setVolume(int volume) {
         this.volume = volume;
     }
 }
